variables_entrypoint: enforcing locale for all base code
This script is called at the early stage.
Some of the code like in zap_devices, consider the locale to be english.

The code does something like "parted -s /dev/sda print | grep -sq boot" to insure not killing a root device.
In a french locale as mine, the grep doesn't work and could kill my host.

Most of the systems runs in english locale but enforcing LC_ALL=C at the
early stage guarantee a reproducible environment for all systems/env.