common_functions: expliciting arguments for dev_part()
The dev_part function is using ${1} and ${2} in the whole function while
never really mention what are those two arguments.

This simple patch is just about naming the arguments at the beginning of
the script and then use it.

This patch is purely cosmetic but makes the code reading easier for new
contributors.