osd_disk_prepare: zapping weird devices only with OSD_FORCE_ZAP
The actual code is dangerous because if parted return an error, the disk is immediately zapped without a clear consent.
That's very dangerous as parted can return an error for any reason.
As parted doesn't classify errors in its return code, a script cannot get how critical is the error. It's up to a human to do so.
Making a mistake here could be destroying user's data without a single warn. That's seriously brutal.

This patch is implementing a more {user,data}-friendly approach.

- Before zapping a device, it's existence should be tested.
- Then, testing if the disk is consistent :
  - if not, we inform the user verify his disk and tell user to enable OSD_FORCE_ZAP to really destroy it
  - if yes, let's continue to search for exisiting ceph metadata on it